Changes to adding/removing entities/components.

+++ Source/ECSL/Framework/Components/Tables/DataMap.cpp
@@ -24,8 +24,7 @@ namespace ECSL
 			for (auto it = m_dataMap->begin(); it != m_dataMap->end(); ++it)
 				delete it->second;
 			delete(m_dataMap);
-		}
-			
+		}	
 	}
 	void DataMap::ClearTable()
 	{
@@ -41,31 +40,34 @@ namespace ECSL
 	{
 		return (*m_dataMap)[_id];
 	}
+
 	DataLocation DataMap::GetData(unsigned const int _id, unsigned const int _index) const
 	{
 		return (*m_dataMap)[_id] + _index;
 	}
+
 	void DataMap::SetData(unsigned const int _id, void* _data, unsigned const int _byteCount)
 	{
 		memcpy((*m_dataMap)[_id], _data, _byteCount);
 	}
+
 	void DataMap::SetData(unsigned const int _id, unsigned const int _index, void* _data, unsigned const int _byteCount)
 	{
 		memcpy((*m_dataMap)[_id] + _index, _data, _byteCount);
 	}
 
-	const int DataMap::GetRowCount() const
+	const unsigned int DataMap::GetRowCount() const
 	{
-		return m_dataMap->size();
+		return (unsigned int)m_dataMap->size();
 	}
 
-	const int DataMap::GetBytesPerRow() const
+	const unsigned int DataMap::GetBytesPerRow() const
 	{
 		return m_bytesPerRow;
 	}
 
-	const int DataMap::GetMemoryAllocated() const
+	const unsigned int DataMap::GetMemoryAllocated() const
 	{
-		return sizeof(DataMap) + m_bytesPerRow * m_dataMap->size();
+		return sizeof(DataMap) + m_bytesPerRow * (unsigned int)m_dataMap->size();
 	}
 }
