+++ Source/ECSL/Managers/ComponentTypeManager.h
@@ -21,7 +21,7 @@ namespace ECSL
 		void LoadComponentTypesFromFile(const std::string& _filePath);
 
 		ComponentType* GetComponentType(int _componentTypeId);
-		unsigned int GetComponentTypeCount(){ return m_componentTypes->size(); }
+		unsigned int GetComponentTypeCount(){ return (unsigned int)m_componentTypes->size(); }
 
 		template<typename ComponentType>
 		unsigned int GetTableId();
@@ -32,7 +32,7 @@ namespace ECSL
 		int m_nextTableId;
 		Parser* m_parser;
 		ComponentTypeReader* m_componentTypeReader;
-		std::map<int, ComponentType*>* m_componentTypes;
+		std::map<unsigned int, ComponentType*>* m_componentTypes;
 		std::unordered_map<std::string, unsigned int>* m_stringTableId;
 
 		template<typename ComponentType>
